Check if data objects exist in a list of Data Packages.

This function is primarily intended to assist Mark Schildhauer in preparation for the Carbon synthesis working group.

data_objects_exist(mn, pids, write_to_csv = FALSE, folder_path = NULL,
  file_name = NULL)

Arguments

mn
(MNode/CNode) The Node to query for Object sizes

pids
(character) The identifier of the Data Packages' Metadata

write_to_csv
(logical) Optional. Write the query results to a csv

folder_path
(character) Optional. Folder to write results to

file_name
(character) Optional. Name of results file in csv format

Value

(data.frame) Data frame containing query results. TODO check if formatting of pid is url then format