Download one or multiple Data Packages

This function is wrapper for download_package It downloads all of the Data Objects in a Data Package to the local filesystem. It is particularly useful when a Data Package is too large to download using the web interface.

download_packages(mn, resource_map_pids, ...)

Arguments

mn
(MNode) The Member Node to download from.

resource_map_pids
(chraracter) The identifiers of the Resource Maps for the packages to download.

...
Allows arguments from download_package

Details

Setting check_download_size to TRUE is recommended if you are uncertain of the total download size and want to avoid downloading very large Data Packages.

This function will also download any data objects it finds in any child Data Packages of the input data package. If you would only like to download data from one Data Package, set download_child_packages to FALSE.

Examples

# NOT RUN {
cn <- CNode("PROD")
mn <- getMNode(cn, "urn:node:ARCTIC")
download_packages(mn, c("resource_map_doi:10.18739/A21G1P" "resource_map_doi:10.18739/A2RZ6X"),
check_download_size = FALSE, prefix_file_names = TRUE, convert_excel_to_csv == TRUE)
# }