Edit a single attribute.

This function edits the slots of a single attribute in an existing attribute table.

edit_attribute(eml, dataTableNumber, attributeNumber,
  attributeName = NULL, attributeDefinition = NULL, domain = NULL,
  measurementScale = NULL, unit = NULL, numberType = NULL,
  definition = NULL, formatString = NULL, missingValueCode = NULL,
  missingValueCodeExplanation = NULL)

Arguments

eml
The eml object containing the attributeList

dataTableNumber
The number of the dataTable containing the attributeList; in other words its position in the dataTable list.

attributeNumber
The number of the attribute of interest; in other words its position in the attributeList.

attributeName
The new name to give to the attribute

attributeDefinition
The new attributeDefinition to give to the attribute

domain
The new domain to give to the attribute

measurementScale
The new measurementScale to give to the attribute

unit
The new unit to give to the attribute

numberType
The new numberType to give to the attribute

definition
The new definition (for textDomain) to give to the attribute

formatString
The new formatString to give to the attribute

missingValueCode
The new missing value code to give to the attribute

missingValueCodeExplanation
The new missing value code explanation to give to the attribute

Details

Can only be used on attributes entirely defined within the "attributes" slot of attributeList; it cannot be used to edit the factor table of an enumeratedDomain. In cases with very large attribute lists, user may want to use which_in_eml function first to locate the attribute position.

Examples

# NOT RUN {
#To change an attribute's name and add a missing value code where there hadn't been one:
new_eml <- edit_attribute(eml, 2, 8, attributeName = "new name", missingValueCode = "NA",
                          missingValueCodeExplanation = "Data unavailable")
# To change an attribute's attributeDefinition, as well as change measurementScale
# from nominal to ratio (requires also changing domain, unit, and numberType, as well
# as setting definition to NA):
new_eml <- edit_attribute(eml, 1, 2, attributeDefinition = "new definition",
domain = "numericDomain", measurementScale = "ratio", unit = "dimensionless",
numberType = "whole", definition = NA)
# }