Get all attribute from an EML document. Document for the function get_attributes from the EML(url) Package

Get all attribute from an EML document. Document for the function get_attributes from the EML(url) Package

get_meta_attributes(node = "ADC", metadata)

Arguments

node
(MNode/CNode) The Dataone Node that stores the Metadata object, from https://cn.dataone.org/cn/v2/node

metadata
(S4 / character) Optional. Either the full eml S4 object document or the url of the Dataone Package. The S4 object input is a more reliable method.

Value

(list) A list of all attribute metadata from the EML in data.frame objects

Examples

# NOT RUN {
node <- dataone::getMNode(cn, 'urn:node:ARCTIC')
eml <- EML::read_eml(rawToChar(dataone::getObject(node, "doi:10.18739/A23W02")))
attributes <- datamgmt::get_meta_attributes(metadata = eml)

attributes <- datamgmt::get_meta_attributes("ADC", "https://arcticdata.io/catalog/#view/doi:10.18739/A23W02")

# switch nodes
cn <- dataone::CNode('PROD')
knb <- dataone::getMNode(cn,"urn:node:KNB")
attributes <- get_meta_attributes("KNB", "https://knb.ecoinformatics.org/#view/doi:10.5063/F1639MWV")
# }