Guess the Member Node

Guess the member node that stores a Dataone object based on its unique identifier (pid) and coordinating node (cn). In most cases the object is stored on the Production ("PROD") Node, however this function can search across all coordinating nodes. If only one member node is identified this function returns the member node as a Dataone "MNode" object. If multiple member nodes are identified a vector of nodes is printed.

guess_member_node(cn = "PROD", pid)

Arguments

cn
(character) A character vector of coordinate nodes to search. Defaults to "PROD". Can be set to any combination of ("PROD", "STAGING", "STAGING2", "SANDBOX", "SANDBOX2", "DEV", "DEV2").

pid
(character) The Datone unique object identifier.

Examples

# NOT RUN {
The following two calls are equivalent:
mn <- guess_member_node("doi:10.18739/A2G287")
mn <- guess_member_node("PROD", "doi:10.18739/A2G287")

Search all coordinating nodes:
cn = c("PROD", "STAGING", "STAGING2", "SANDBOX", "SANDBOX2", "DEV", "DEV2"))
mn <- guess_member_node(cn, "doi:10.18739/A2G287")
# }