Check congruence of attributes and data for a given dataset and dataTable

This function is called by qa_package but can be used on its own to test congruence between a dataTable and a data object (data.frame). See qa_package documentation for more details.

qa_attributes(dataTable, data, checkEnumeratedDomains = TRUE)

Arguments

dataTable
(dataTable) EML dataTable, otherEntity, or spatialVector associated with the data object.

data
(data.frame) Data frame of data object.

checkEnumeratedDomains
(logical) Default TRUE. Compare unique values in data to defined enumerated domains in EML.

Details

Purpose: QA function to check that attributes match values in the data

Functions:

Names: Check that column names in attributes match column names in data frame. Possible conditions to check for:

attributeList does not exist for data frame

Physical has not been set correctly

Some of the attributes that exist in the data do not exist in the attributeList

Some of the attributes that exist in the attributeList do not exist in the data

Typos in attribute or column names resulting in nonmatches

Domains: Check that attribute types match attribute types in data frame. Possible conditions to check for:

nominal, ordinal, integer, ratio, dateTime

If domain is enumerated domain, enumerated values in the data are accounted for in the enumerated definition

If domain is enumerated domain, enumerated values in the enumerated definition are all represented in the data

Type of data does not match attribute type

Values: Check for accidental characters in the data frame (one character in a column of integers)

Examples

# NOT RUN {
# Run attribute checks on a data.frame and its associated EML dataTable

qa_attributes(dataTable, dataObject)
# }