Check package including congruence of attributes and data

This function assumes correctness in the resource map and data. Purpose: QA script to check a DataONE EML package. Checks that attributes match values in the data. Note: this function also calls qa_attributes, qa_creator_ORCIDs, and qa_access.

qa_package(node, pid, readAllData = TRUE, check_attributes = TRUE,
  check_creators = FALSE, check_access = FALSE)

Arguments

node
(MNode) Member Node where the PID is located.

pid
(character) The PID of a resource map.

readAllData
(logical) Default TRUE. Read all data from remote and check that column types match attributes, otherwise only pull first 10 rows. Only applicable to public packages (private packages will read complete dataset). If check_attributes = FALSE, no rows will be read.

check_attributes
(logical) Default TRUE. Calls qa_attributes. Checks congruence of attributes and data.

check_creators
(logical) Default FALSE. Calls qa_creator_ORCIDs. Checks if each creator has an ORCID. Will also run if check_access = TRUE.

check_access
(logigal) Default FALSE. Calls qa_access. Checks if each creator has full access to the metadata, resource map, and data objects. Will not run if the checks associated with check_creators fail.

Value

invisible

Examples

# NOT RUN {
# Run QA checks

qa_package(mn, pid, readAllData = FALSE, check_attributes = TRUE,
           check_creators = FALSE, check_access = FALSE)
# }