Download multiple data packages

This function is a convenience wrapper for download_package() when downloading multiple data packages. It downloads all of the data objects in a data package to the local filesystem. It is particularly useful when a data package is too large to download using the web interface.

download_packages(mn, resource_map_pids, download_directory, ...)

Arguments

mn
(MNode) The Member Node to download from.

resource_map_pids
(chraracter) The PIDs of the resource maps for the packages to download.

download_directory
(character) The path of the directory to download the packages to.

...
Allows arguments from download_package().

Details

Setting check_download_size to TRUE is recommended if you are uncertain of the total download size and want to avoid downloading very large data packages.

This function will also download any data objects it finds in any child data packages of the input data package. If you would only like to download data from one data package, set download_child_packages to FALSE.

See also

Examples

# NOT RUN {
cn <- CNode("PROD")
mn <- getMNode(cn, "urn:node:ARCTIC")

download_packages(mn, c("resource_map_doi:10.18739/A21G1P", "resource_map_doi:10.18739/A2RZ6X"),
"/home/dmullen/downloads", prefix_file_names = TRUE, download_column_metadata = TRUE,
convert_excel_to_csv = TRUE)
# }