Guess the Member Node

Guess the Member Node that stores a DataONE object based on its unique identifier (pid) and coordinating node (cn). In most cases the object is stored on the Production ("PROD") Node, however this function can search across all coordinating nodes. If only one Member Node is identified this function returns the Member Node as a DataONE "MNode" object. If multiple Member Nodes are identified a vector of nodes is printed.

guess_member_node(pid, cn = "PROD")

Arguments

pid
(character) The DataONE unique object identifier. A DataONE package URL can also be used as input (although this method is less reliable).

cn
(character) A character vector of coordinate nodes to search. Defaults to "PROD". Can be set to any combination of ("PROD", "STAGING", "STAGING2", "SANDBOX", "SANDBOX2", "DEV", "DEV2").

Examples

# NOT RUN {
The following two calls are equivalent:
mn <- guess_member_node("doi:10.18739/A2G287")
mn <- guess_member_node("doi:10.18739/A2G287", "PROD")

Use a DataONE package URL
mn <- guess_member_node("https://arcticdata.io/catalog/#view/doi:10.18739/A2TX35587")
Search all coordinating nodes:
cn = c("PROD", "STAGING", "STAGING2", "SANDBOX", "SANDBOX2", "DEV", "DEV2")
mn <- guess_member_node("doi:10.18739/A2G287", cn)
# }